extract logic out into AppChromeMenu
